# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Freightrek Server API",
    "version": "1.0.0",
    "description": "REST API for Freightrek freight management system with admin authentication and RBAC",
    "contact": {
      "name": "Freightrek Team",
      "email": "support@freightrek.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "https://freightrekapi.vercel.app",
      "description": "Production (Vercel)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development (local)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token obtained from login endpoint"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/admin/agency/invoice/order/{orderId}": {
      "get": {
        "summary": "The invoice for one of this agency's orders",
        "tags": [
          "Invoice - Agency"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice details"
          },
          "403": {
            "description": "Invoice belongs to another agency"
          }
        }
      }
    },
...
    "/admin/branch/parcel-order": {
      "post": {
        "summary": "Book a parcel order (branch / franchise login)",
        "tags": [
          "Parcel Flow - Branch"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "bookingCustomer",
                  "paymentType",
                  "deliveryCustomer",
                  "parcelDetails"
                ],
                "properties": {
                  "bookingCustomer": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "mobileNumber": {
                        "type": "string"
                      }
                    }
                  },
                  "paymentType": {
                    "type": "string",
                    "enum": [
                      "Paid",
                      "To Pay",
                      "Credit"
                    ]
                  },
                  "deliveryCustomer": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "mobileNumber": {
                        "type": "string"
                      },
                      "deliveryAgency": {
                        "type": "string",
                        "description": "Optional. Destination agency ObjectId, picked from GET /admin/agency/parcel-order/options/delivery-branches. Can be set later via PUT /admin/agency/parcel-order/{{id}}.\n"
                      },
                      "deliveryBranch": {
                        "type": "string",
                        "description": "Deprecated alias for deliveryAgency"
                      }
                    }
                  },
                  "pickupAddress": {
                    "type": "string",
                    "description": "Where the parcel is collected from the booking customer"
                  },
                  "deliveryAddress": {
                    "type": "string",
                    "description": "Where the parcel is delivered to the delivery customer"
                  },
                  "parcelDetails": {
                    "type": "object",
                    "properties": {
                      "article": {
                        "type": "string"
                      },
                      "remarks": {
                        "type": "string"
                      },
                      "numberOfParcels": {
                        "type": "integer"
                      },
                      "approximateValue": {
                        "type": "number"
                      }
                    }
                  },
                  "transportationCharge": {
                    "type": "number"
                  },
                  "waybill": {
                    "type": "string",
                    "description": "Optional. Carrier waybill / AWB number"
                  },
                  "vehicleType": {
                    "type": "string",
                    "description": "Optional. Vehicle class booked for the movement"
                  },
                  "vehicleCapacity": {
                    "type": "string",
                    "description": "Optional. Capacity of that vehicle (e.g. \"2 Ton\")"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order booked with status \"Order Created\""
          },
          "403": {
            "description": "Branch access required"
          }
        }
      },
      "get": {
        "summary": "List this branch's parcel orders",
        "tags": [
          "Parcel Flow - Branch"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hubAssignment",
            "schema": {
              "type": "string",
              "enum": [
                "assigned",
                "unassigned"
              ]
            }
          },
          {
            "in": "query",
            "name": "direction",
            "schema": {
              "type": "string",
              "enum": [
                "outgoing",
                "incoming"
              ]
            },
            "description": "outgoing = booked at this branch, incoming = addressed to this branch for delivery. Omit for both.\n"
          },
          {
            "in": "query",
            "name": "paymentType",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orders booked at or addressed to this branch"
          }
        }
      }
    },
    "/admin/branch/parcel-order/outward": {
      "get": {
        "summary": "This branch's outward register",
        "description": "Parcels booked at this branch and sent out. The branch comes from the token. Same filters as the list; counterpartAgency narrows it to one destination branch.\n",
        "tags": [
          "Parcel Flow - Branch"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "counterpartAgency",
            "schema": {
              "type": "string"
            },
            "description": "Destination branch ObjectId"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "paymentType",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dateFrom",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "dateTo",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ direction: outward, agency, orders, totals, pagination }"
          }
        }
      }
    },
    "/admin/branch/parcel-order/inward": {
      "get": {
        "summary": "This branch's inward register",
        "description": "Parcels booked at another branch and addressed to this one for delivery. counterpartAgency here narrows it to one origin branch.\n",
        "tags": [
          "Parcel Flow - Branch"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "counterpartAgency",
            "schema": {
              "type": "string"
            },
            "description": "Origin (booking) branch ObjectId"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "paymentType",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dateFrom",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "dateTo",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ direction: inward, agency, orders, totals, pagination }"
          }
        }
      }
    },
...
    "/admin/branch/parcel-order/wallet-preview": {
      "get": {
        "summary": "What a booking of this amount will take from the branch wallet",
        "description": "Booking a parcel debits the admin's share of the transportation charge from this branch's wallet. Call this before booking to show the split and warn on a low balance.\n",
        "tags": [
          "Parcel Flow - Branch"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "amount",
            "required": true,
            "schema": {
              "type": "number",
              "example": 200
            },
            "description": "The transportation charge being quoted"
          }
        ],
        "responses": {
          "200": {
            "description": "orderAmount, profitPercentage, branchProfitAmount, adminShareAmount, walletDebitAmount (the full order amount taken from the wallet), walletBalance, balanceAfterBooking, sufficientBalance\n"
          },
          "400": {
            "description": "A valid amount is required"
          }
        }
      }
    },
    "/admin/branch/parcel-order/{id}": {
      "get": {
        "summary": "Get one of this branch's parcel orders",
        "tags": [
          "Parcel Flow - Branch"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details (branch/hub populated)"
          },
          "403": {
            "description": "Order belongs to another branch"
          }
        }
      },
      "put": {
        "summary": "Update booking details of this branch's order",
        "tags": [
          "Parcel Flow - Branch"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          },
          "403": {
            "description": "Order belongs to another branch"
          }
        }
      }
    },
    "/admin/branch/parcel-order/{id}/tracking": {
      "get": {
        "summary": "Status timeline of this branch's order",
        "tags": [
          "Parcel Flow - Branch"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current status + timeline"
          }
        }
      }
    },
    "/admin/branch/parcel-order/{id}/charge": {
      "patch": {
        "summary": "Update transportation charge on this branch's order",
        "tags": [
          "Parcel Flow - Branch"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "transportationCharge"
                ],
                "properties": {
                  "transportationCharge": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Charge updated"
          }
        }
      }
    },
    "/admin/branch/parcel-order/{id}/status": {
      "patch": {
        "summary": "Branch-side status update",
        "description": "A branch may set only its own stages — Parcel Collected, Parcel Dispatched, Parcel Arrived at Branch, Parcel Received at Branch, Delivered. The lifecycle moves forward only.\n",
        "tags": [
          "Parcel Flow - Branch"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "Parcel Collected",
                      "Parcel Dispatched",
                      "Parcel Arrived at Branch",
                      "Parcel Received at Branch",
                      "Delivered"
                    ]
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          },
          "400": {
            "description": "Backwards move",
            "or hub stage without a hub": null
          },
          "403": {
            "description": "Status not allowed for a branch"
          }
        }
      }
    },
...
    "/admin/invoice/order/{orderId}": {
      "get": {
        "summary": "The invoice raised for a parcel order",
        "tags": [
          "Invoice"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice details"
          },
          "404": {
            "description": "No invoice for this order"
          }
        }
      }
    },
...
    "/admin/parcel-order/inward": {
      "get": {
        "summary": "Inward register - parcels addressed to an agency for delivery",
        "description": "Same shape as the outward register; counterpartAgency here filters the origin (booking) agency.\n",
        "tags": [
          "Parcel Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "agency",
            "schema": {
              "type": "string"
            },
            "description": "Agency whose register this is. Required for admin, ignored for an agency login"
          },
          {
            "in": "query",
            "name": "counterpartAgency",
            "schema": {
              "type": "string"
            },
            "description": "Origin agency, to see only what came from one agency"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "paymentType",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dateFrom",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "dateTo",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ direction: inward, agency, orders, totals, pagination }"
          },
          "400": {
            "description": "agency is required for an admin login"
          },
          "403": {
            "description": "Not available to a hub login"
          }
        }
      }
    },
...
    "/admin/parcel-order/{id}/tracking": {
      "get": {
        "summary": "Full status timeline for a parcel order",
        "tags": [
          "Parcel Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current status + statusHistory timeline"
          }
        }
      }
    },
...
  },
...
}